+++ Cosette/Engine/Ai/Transposition/TranspositionTable.cs
@@ -9,11 +9,11 @@ namespace Cosette.Engine.Ai.Transposition
         private static TranspositionTableEntry[] _table;
         private static ulong _size;
 
-        public static unsafe void Init(ulong sizeMegabytes)
+        public static unsafe void Init(int sizeMegabytes)
         {
             var entrySize = sizeof(TranspositionTableEntry);
 
-            _size = sizeMegabytes * 1024ul * 1024ul / (ulong)entrySize;
+            _size = (ulong)sizeMegabytes * 1024ul * 1024ul / (ulong)entrySize;
             _table = new TranspositionTableEntry[_size];
         }
